The option delete is available again using DTO

+++ back/src/main/java/co/com/sofka/crud/repositories/TodoRepository.java
@@ -4,10 +4,11 @@ import co.com.sofka.crud.dto.TodoDTO;
 import co.com.sofka.crud.entities.Todo;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
+import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 
-public interface TodoRepository extends CrudRepository<Todo, Long> {
+public interface TodoRepository extends CrudRepository<Todo, Integer> {
     String queryAll = "select new co.com.sofka.crud.dto.TodoDTO(t.id, t.name, t.completed, t.groupListId)"
             + " from Todo t"
             + " where 1=1";
@@ -17,4 +18,5 @@ public interface TodoRepository extends CrudRepository<Todo, Long> {
 
     @Query(value = queryAll + "and t.id=?1")
     public List<TodoDTO> getTodoById(int id);
+
 }
